+++ packages/graphql-codegen-core/src/types.ts
@@ -285,5 +285,6 @@ export function isCustomProcessingFunction(
 }
 
 export function isGeneratorConfig(config: GeneratorConfig | CustomProcessingFunction): config is GeneratorConfig {
+  console.log(config);
   return typeof config !== 'function' && !!(config as any).inputType;
 }